+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## v0.2.10 - 2017-05-19
+### Added:
+ - Search in resume black/white lists
+ - Active/archived/hidden vacancies
+ - Hide/restore a vacancy
+
+
 ## v0.2.9 - 2017-05-13
 ### Added:
  - Ability to change token, without creating a new instance.